+++ invoice/src/main/java/az/cybernet/invoice/dto/response/InvoiceResponse.java
@@ -4,6 +4,7 @@ import az.cybernet.invoice.enums.Status;
 import lombok.*;
 import lombok.experimental.FieldDefaults;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
@@ -19,7 +20,7 @@ public class InvoiceResponse {
     UUID senderId;
     UUID customerId;
     Status status;
-    Double total;
+    BigDecimal total;
     LocalDateTime createdAt;
     LocalDateTime updatedAt;
     String comment;
